correção para o Context da aplicação (context é o que permite que os dados sejam passados para dos os componentes da aplicação
